requirements.txt
streamlit==1.28.0
google-auth-oauthlib==1.0.0
google-auth==2.23.0
google-api-python-client==2.100.0
python-dotenv==1.0.0
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
streamlit-google-auth==0.2.3
sqlite3